libese: initial commit
libese is a low level library for enabling embedded secure
element use from Java or embedded usecases.

This commit defines the client API, the hardware abstraction API,
and supplies functional implementation of the T=1 protocol (frequently
used as the wire protocol for SPI attached smart cards/eSEs).

Included in this is one hardware reference implementation using
a NXP PN80T developer board wired to a HiKey 6220.

Test: baseline unittests; manual run of the examples/ese_nxp_sample; use
of relay with hardware
Bug: 34193615

Change-Id: I98037793bc29b3730c5301ee9cd5e6cd7465117d