libese-hw: clean up reference nxp cooldown
While this code should move into platform-specific code
(kernel, spidev helper), it is going in common first such
that the platform will not power-down a chip that has
an elevated attack counter.  In practice, the system needs
to keep the eSE powered and unused for the wait time
(in minutes). That means it really should be at the platform
pinch-points.

Change-Id: I1c1c795055324cc80564f4246199de65775a0603
Bug: 36778596
Test: works on sample hardware