libese/*: clarify cooldown comments
This change cleans up comments and documentation around libese close
behavior and chip cooldown.

libese supports keeping an eSE powered on after a session is completed
to mirror the behavior of TPM dictionary attack defense (DAD) "cooldown"
implemented in external/trunks and
https://chromium.googlesource.com/chromiumos/platform2/+/master/cryptohome
for working with TCG TPM 1.2 and 2.0 spec chips.

On Chromebooks, handling this case was critical for assuring access to
secure boot storage information and sign in keys.   As eSEs adhere to a
different set of standards (GP) which do not include the same scope as
the TCG specs as well as a dynamic power lifecycle (not always powered),
they may need to remain powered for any number of
implementation-specific reasons.  The prior changes had assumed details
which were not accurate, and the documentation and code is now
up-to-date.

Test: booted and works as expected.  ese-boot-tool with a minor fix
      validates against avb_storage.cap.
Bug: 36778596
Change-Id: I6584f6b8d2b29cb58ea93c9bf619d5fb84f60596