libese-teq1: fix S(RESYNC, REQUEST) retransmission
When a device is non-response, T=1 will send
resynchronization requests.  After three attempts
it should either warm reset the device or deactivate.

The error path was not checking if the triggering
transmission was a resync requests and was clobbering
the error and retransmission counters resulting in
infinite re-attempts.

Test: replace/dev/pn81a with /dev/zero and ensure that it eventually
times out and returns a hard failure.
Bug: none
Change-Id: Iac512bb37f7f04deefe91570547840a7704542d9