esed: don't use HW binder status.
Previously, this would return EX_ILLEGAL_ARGUMENT with an error message
for debugging. Now, it returns FAILED and logs the reason. A new status
code is not needed as clients of the HAL don't need to distinguish this
cause of failure in the usual case and the log has the information
needed for debugging.

Change-Id: I30587f0bf3e2776987de39a0577a88131533fb46
Test: manual -- enroll and authenticate using Weaver
Fix: 38163768