Weaver: add command to erase slot value
Only the value is erased; the key remains unchanged. This allows the
secret to be destroyed while stil allowing read requests to be validated
against the key.

This can be used for FRP as the credential still needs to be validated
by Weaver but we want the secret to be wiped.

Test: libese-app-weaver-test
Bug: 36814845
Change-Id: I01843a4c51dfbb0af77d6076bfd684f704901b79