libese-teq1: global resync maximum
Set a global maximum for software resyncs before powering the device
on and off.

This also implies a global maximum of attempts as only one physical
reset is allowed prior to a hard reset.

(Note, this pulls in the fake hw hack from libese-boot-app which
 should be migrated to libese-cpp in the future as libese-boot is
 likely to go away.)

Bug: 63546784
Test: boots; normal functions work; teq1_unittests with the error case
passes.
Merged-In: I403bd90875c08372f043d765497640c216592216

Change-Id: I394b7be729bc714d7b19c2ce9af556c8be251f68